+++ NOBY/NOBY.Api/Endpoints/Cases/GetTaskList/Dto/WorkflowTask.cs
@@ -2,9 +2,9 @@
 
 public sealed class WorkflowTask
 {
-    public int TaskId { get; set; }
+    public long TaskId { get; set; }
 
-    public int TaskProcessId { get; set; }
+    public long TaskProcessId { get; set; }
 
     public int TypeId { get; set; }
